test: fix imports due to refactoring of jobs
