+++ test/run/run.c.spec.ts
@@ -1,6 +1,7 @@
 import {execute} from '../../src/tasks/'
 import {expect} from 'chai'
-import { RunJob } from '../../src/tasks/job'
+import {RunJob} from '../../src/tasks/jobs/run'
+
 describe('run - c', () => {
   it('.c file runs correctly', async () => {
     const runResult = await execute(new RunJob({
