+++ test/scenarios/projectScenario.spec.ts
@@ -1,16 +1,16 @@
 import { expect } from 'chai'
-import { ProjectJob } from '../../src/tasks/job'
+import { ProjectJob } from '../../src/tasks/jobs/project';
 import ProjectScenarion from '../../src/tasks/scenarios/project'
 
 describe('Project Scenario', () => {
   it('should setup', async () => {
-    const job: ProjectJob = {
-      id: 1,
-      source: '',
-      lang: 'nodejs',
-      timelimit: 20,
-      scenario: 'problem',
-      problem: ''
-    }
+    // const job: ProjectJob = {
+    //   id: 1,
+    //   source: '',
+    //   lang: 'nodejs',
+    //   timelimit: 20,
+    //   scenario: 'problem',
+    //   problem: ''
+    // }
   })
 })
